requirements.txt
python-telegram-bot==12.8
requests==2.23.0
python==3.8